Reorganize traverse functions
