+++ src/core/renderer/index.js
@@ -9,3 +9,4 @@ export * from './constants.js';
 
 export { LRUCache } from './utilities/LRUCache.js';
 export { PriorityQueue } from './utilities/PriorityQueue.js';
+export * as TraverseUtils from './utilities/TraversalUtils.js';
